+++ tests/test_reprs.py
@@ -1,15 +1,16 @@
 import ee
 
 import eerepr
-from eerepr.repr import _repr_html_
 
 
 def test_full_repr(data_regression):
     """Regression test the full HTML repr (with CSS and JS) of a nested EE object."""
     from tests.test_html import get_test_objects
 
-    objects = get_test_objects().items()
-    rendered = _repr_html_(ee.List([obj[1] for obj in objects]))
+    eerepr.initialize()
+
+    objects = get_test_objects().values()
+    rendered = ee.List([obj for obj in objects])._repr_html_()
     data_regression.check(rendered)
 
 
@@ -38,3 +39,19 @@ def test_existing_repr_html():
     # reset shouldn't remove the existing repr
     eerepr.reset()
     assert obj._repr_html_() == "foo"
+
+
+def test_scripts_sanitized():
+    """Test that scripts within objects are escaped."""
+    eerepr.initialize()
+
+    script_injection = "<script>alert('foo')</script>"
+
+    obj = ee.String(script_injection)
+    assert "<script>" not in obj._repr_html_()
+
+    obj = ee.List([script_injection])
+    assert "<script>" not in obj._repr_html_()
+
+    obj = ee.Dictionary({script_injection: script_injection, "type": script_injection})
+    assert "<script>" not in obj._repr_html_()