+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 Nothing yet.
 
-## [0.1.0] - 2024-12-06
+## [0.1.0] - 2025-01-10
 
 ### Breaking Changes